+++ src/ServiceAccountJwtAccessCredentials.php
@@ -17,11 +17,6 @@
 
 namespace Google\Auth;
 
-use GuzzleHttp\ClientInterface;
-use GuzzleHttp\Client;
-use GuzzleHttp\Stream\Stream;
-use GuzzleHttp\Exception\ClientException;
-use GuzzleHttp\Exception\ServerException;
 use Http\Client\HttpClient;
 
 /**
